container-applications/IoT/requirements.txt
click
dominate
Flask
itsdangerous
Jinja2
MarkupSafe
onetimepass
six
visitor
pymysql
Werkzeug
gunicorn
google-api-core
google-cloud-datastore
google-api-python-client
google-cloud-core
google-cloud-storage
google-cloud-runtimeconfig
google-cloud-pubsub
google-cloud-iot
requests